add post sharing feature as well inspired by facebook where  "Clicking the share button on a post will create a card with the post to be shared in the create post modal and user shall be able to either add there thaughts/captions or share it without that"
also

"When a user is viewing a post that contains a shared post:

If the original shared post is already in the DOM:

Clicking on the shared post container should smoothly scroll the user to the original post's location in the DOM.

If the original shared post is not in the DOM:

Clicking the shared post container should trigger fetching of the original post data.

Once fetched, the original post should be dynamically inserted directly below the current post being viewed.

Then, smoothly scroll the user to the newly inserted original post."

also change the mockdata to add so many posts including orignal user's own and shared posts